+++ common/src/main/java/net/conczin/immersive_gateways/data/PortalDataManager.java
@@ -82,7 +82,7 @@ public class PortalDataManager {
                 );
 
                 // Real resolved position
-                BlockPos realTarget = placeStructure(level, target, attempt >= SEARCH_ATTEMPTS - SEARCH_FALLBACK_ATTEMPTS, attempt != SEARCH_ATTEMPTS - 1);
+                BlockPos realTarget = placeStructure(level, target, attempt >= SEARCH_ATTEMPTS - SEARCH_FALLBACK_ATTEMPTS, attempt != SEARCH_ATTEMPTS - 1, attempt != SEARCH_ATTEMPTS - 1);
                 if (realTarget != null) {
                     target = realTarget;
                     break;
@@ -117,9 +117,14 @@ public class PortalDataManager {
         state.add(pair);
     }
 
-    public static BlockPos placeStructure(ServerLevel level, BlockPos pos, boolean useFallback, boolean checkInhabitedTime) {
+    public static BlockPos placeStructure(ServerLevel level, BlockPos pos, boolean useFallback, boolean checkInhabitedTime, boolean checkWorldBorder) {
         Registry<Structure> registry = level.registryAccess().registry(Registries.STRUCTURE).orElse(null);
         if (registry == null) {
+            return null;
+        }
+
+        // Prevent generating outside the world border
+        if (checkWorldBorder && !level.getWorldBorder().isWithinBounds(pos)) {
             return null;
         }