+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.0.6
+
+* Prevent gateways outside the world border
+
 # 0.0.5
 
 * Fixed issues with config loading